requirements.txt
playwright>=1.40.0
apscheduler>=3.10.0
pyyaml>=6.0
